+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## HEAD
+
+* Deprecate `:concurrency` option in S3 storage (@janko)
+
+* Don't use a thread pool in `S3#concatenate` (@janko)
+
 ## 2.3.0 (2019-05-14)
 
 * Allow uploading files larger than 50 GB to S3 storage by scaling the part size according to upload length (@janko)